requirements.txt
appdirs==1.4.2
click==6.7
coreapi==2.2.4
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.10.5
djangorestframework==3.5.4
drfdocs==0.0.11
gunicorn==19.7.0
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
openapi-codec==1.3.1
packaging==16.8
pyaml==16.12.2
Pygments==2.2.0
pyparsing==2.1.10
PyYAML==3.12
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
